feat(frontend): move sign-out into account settings
Keep the sidebar user block to support/admin actions, open settings from a gear next to the email like the org selector, and put logout on the account settings page.
